requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy 
joblib
numpy 
pandas
seaborn
matplotlib
scikit-learn
authlib
passlib
bcrypt==4.0.1
pydantic-settings
pytest
httpx
redis
redis[async]
snakeviz
line_profiler
locust
prometheus-fastapi-instrumentator
ipykernel
xgboost
python-jose[cryptography]
psycopg2-binary

